chore: fix warnings of terraform 0.12.20
When upgrading to terraform 0.12.20 it produces warnings about quotes
around types and interpolation with only one value inside. This resolves
them, so terraform validate doesn't produce any warnings on 0.12.20

I also cleaned up the note about terraform 0.11, because we're using
features that are not supported in 0.11 (e.g. `list(string)`)